refactor(cli): move dataset formatting to _format
